VIX-3044 Align xModel import patterns
Custom and circle xModel imports had drifted into different naming and organization patterns, making future model types harder to follow.

Rename the shared import model, split circle parsing and generation responsibilities, centralize common xModel attributes, and move prop assembly behind a common assembler.
