+++ src/Vixen.Modules/App/CustomPropEditor/Import/XLights/XModelCommonConfiguration.cs
@@ -1,0 +1,15 @@
+namespace VixenModules.App.CustomPropEditor.Import.XLights
+{
+	internal sealed class XModelCommonConfiguration
+	{
+		public string Name { get; init; }
+
+		public int PixelSize { get; init; }
+
+		public string StringType { get; init; }
+
+		public string StrandNames { get; init; }
+
+		public string NodeNames { get; init; }
+	}
+}
